server_worker_api/api/requirements.txt
APScheduler==3.6.0
Flask==1.0.3
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Werkzeug==0.15.4
marshmallow==2.19.2
paho-mqtt==1.4.0
webargs==5.3.1
schedule==0.6.0
gunicorn